# src/infrastructure/api/openapi.json
{
  "components": {
    "schemas": {
      "AlertDetail": {
        "properties": {
          "alert": {
            "$ref": "#/components/schemas/AlertOut"
          },
          "milestone": {
            "$ref": "#/components/schemas/MilestoneOut"
          },
          "patient": {
            "$ref": "#/components/schemas/PatientOut"
          }
        },
        "required": [
          "alert",
          "milestone",
          "patient"
        ],
        "title": "AlertDetail",
        "type": "object"
      },
      "AlertOut": {
        "properties": {
          "actionTaken": {
            "anyOf": [
              {
                "enum": [
                  "guidance",
                  "reschedule",
                  "social_work_referral",
                  "lodging_coordination",
                  "transport_coordination",
                  "other"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Actiontaken"
          },
          "category": {
            "enum": [
              "transport",
              "lodging",
              "financial",
              "schedule",
              "instructions",
              "communication",
              "health_difficulty",
              "other"
            ],
            "title": "Category",
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "title": "Createdat",
            "type": "string"
          },
          "familyContacted": {
            "title": "Familycontacted",
            "type": "boolean"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "internalNote": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Internalnote"
          },
          "milestoneId": {
            "format": "uuid",
            "title": "Milestoneid",
            "type": "string"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "patientId": {
            "format": "uuid",
            "title": "Patientid",
            "type": "string"
          },
          "resolvedAt": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolvedat"
          },
          "risk": {
            "enum": [
              "green",
              "yellow",
              "red"
            ],
            "title": "Risk",
            "type": "string"
          },
          "status": {
            "enum": [
              "open",
              "in_progress",
              "resolved"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "patientId",
          "milestoneId",
          "category",
          "status",
          "risk",
          "familyContacted",
          "createdAt"
        ],
        "title": "AlertOut",
        "type": "object"
      },
...
      "CareTeamPatientRow": {
        "properties": {
          "hasMissed": {
            "title": "Hasmissed",
            "type": "boolean"
          },
          "hasOpenBarrier": {
            "title": "Hasopenbarrier",
            "type": "boolean"
          },
          "nextMilestone": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MilestoneOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "patient": {
            "$ref": "#/components/schemas/PatientOut"
          },
          "reason": {
            "title": "Reason",
            "type": "string"
          }
        },
        "required": [
          "patient",
          "hasOpenBarrier",
          "hasMissed",
          "reason"
        ],
        "title": "CareTeamPatientRow",
        "type": "object"
      },
...
      "ContactFamilyRequest": {
        "properties": {
          "operationId": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Operationid"
          }
        },
        "title": "ContactFamilyRequest",
        "type": "object"
      },
...
      "CreateMilestoneRequest": {
        "properties": {
          "confirmationDeadline": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confirmationdeadline"
          },
          "location": {
            "anyOf": [
              {
                "maxLength": 200,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location"
          },
          "operationId": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Operationid"
          },
          "preparation": {
            "anyOf": [
              {
                "maxLength": 500,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Preparation"
          },
          "scheduledAt": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scheduledat"
          },
          "service": {
            "anyOf": [
              {
                "maxLength": 200,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Service"
          },
          "title": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Title",
            "type": "string"
          },
          "type": {
            "enum": [
              "consultation",
              "laboratory",
              "procedure",
              "treatment",
              "follow_up"
            ],
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title"
        ],
        "title": "CreateMilestoneRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MilestoneOut": {
        "properties": {
          "attendanceConfirmed": {
            "title": "Attendanceconfirmed",
            "type": "boolean"
          },
          "confirmationDeadline": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confirmationdeadline"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "location": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location"
          },
          "patientId": {
            "format": "uuid",
            "title": "Patientid",
            "type": "string"
          },
          "preparation": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Preparation"
          },
          "scheduledAt": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scheduledat"
          },
          "service": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Service"
          },
          "status": {
            "enum": [
              "completed",
              "upcoming",
              "unscheduled",
              "support_needed",
              "rescheduled",
              "missed"
            ],
            "title": "Status",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "type": {
            "enum": [
              "consultation",
              "laboratory",
              "procedure",
              "treatment",
              "follow_up"
            ],
            "title": "Type",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "patientId",
          "type",
          "title",
          "status",
          "attendanceConfirmed",
          "version"
        ],
        "title": "MilestoneOut",
        "type": "object"
      },
...
      "OverviewResponse": {
        "properties": {
          "counts": {
            "$ref": "#/components/schemas/RiskCounts"
          },
          "generatedAt": {
            "format": "date-time",
            "title": "Generatedat",
            "type": "string"
          },
          "openAlerts": {
            "title": "Openalerts",
            "type": "integer"
          }
        },
        "required": [
          "counts",
          "openAlerts",
          "generatedAt"
        ],
        "title": "OverviewResponse",
        "type": "object"
      },
      "PatientOut": {
        "properties": {
          "age": {
            "title": "Age",
            "type": "integer"
          },
          "avatarKey": {
            "title": "Avatarkey",
            "type": "string"
          },
          "caregiverName": {
            "title": "Caregivername",
            "type": "string"
          },
          "contactPhone": {
            "title": "Contactphone",
            "type": "string"
          },
          "displayName": {
            "title": "Displayname",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "operationalRisk": {
            "enum": [
              "green",
              "yellow",
              "red"
            ],
            "title": "Operationalrisk",
            "type": "string"
          },
          "routeStatus": {
            "enum": [
              "on_track",
              "confirmation_needed",
              "support_needed"
            ],
            "title": "Routestatus",
            "type": "string"
          }
        },
        "required": [
          "id",
          "displayName",
          "age",
          "avatarKey",
          "caregiverName",
          "contactPhone",
          "operationalRisk",
          "routeStatus"
        ],
        "title": "PatientOut",
        "type": "object"
      },
...
      "RescheduleMilestoneRequest": {
        "properties": {
          "newScheduledAt": {
            "format": "date-time",
            "title": "Newscheduledat",
            "type": "string"
          },
          "operationId": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Operationid"
          }
        },
        "required": [
          "newScheduledAt"
        ],
        "title": "RescheduleMilestoneRequest",
        "type": "object"
      },
      "ResolveAlertRequest": {
        "properties": {
          "actionTaken": {
            "enum": [
              "guidance",
              "reschedule",
              "social_work_referral",
              "lodging_coordination",
              "transport_coordination",
              "other"
            ],
            "title": "Actiontaken",
            "type": "string"
          },
          "internalNote": {
            "anyOf": [
              {
                "maxLength": 500,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Internalnote"
          },
          "newScheduledAt": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Newscheduledat"
          },
          "operationId": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Operationid"
          }
        },
        "required": [
          "actionTaken"
        ],
        "title": "ResolveAlertRequest",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "API del piloto conectado de Kinti (Fase 2). Datos exclusivamente sint�ticos. El sem�foro representa riesgo operativo de interrupci�n de la ruta, no gravedad cl�nica. No diagnostica, no prescribe y no realiza triaje.",
    "title": "Kinti API",
    "version": "2.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/alerts/{alert_id}": {
      "get": {
        "operationId": "get_alert_api_v1_alerts__alert_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "alert_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Alert Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlertDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Alert",
        "tags": [
          "equipo asistencial"
        ]
      }
    },
    "/api/v1/alerts/{alert_id}/contact": {
      "post": {
        "operationId": "contact_family_api_v1_alerts__alert_id__contact_post",
        "parameters": [
          {
            "in": "path",
            "name": "alert_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Alert Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContactFamilyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlertOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Contact Family",
        "tags": [
          "equipo asistencial"
        ]
      }
    },
    "/api/v1/alerts/{alert_id}/resolve": {
      "post": {
        "operationId": "resolve_alert_api_v1_alerts__alert_id__resolve_post",
        "parameters": [
          {
            "in": "path",
            "name": "alert_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Alert Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResolveAlertRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlertOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Resolve Alert",
        "tags": [
          "equipo asistencial"
        ]
      }
    },
...
    "/api/v1/care-team/alerts": {
      "get": {
        "operationId": "list_alerts_api_v1_care_team_alerts_get",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "pattern": "^(open|in_progress|resolved|pending)$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "risk",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "pattern": "^(green|yellow|red)$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Risk"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AlertOut"
                  },
                  "title": "Response List Alerts Api V1 Care Team Alerts Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Alerts",
        "tags": [
          "equipo asistencial"
        ]
      }
    },
    "/api/v1/care-team/overview": {
      "get": {
        "operationId": "overview_api_v1_care_team_overview_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OverviewResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Overview",
        "tags": [
          "equipo asistencial"
        ]
      }
    },
    "/api/v1/care-team/patients": {
      "get": {
        "operationId": "list_patients_api_v1_care_team_patients_get",
        "parameters": [
          {
            "in": "query",
            "name": "risk",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "pattern": "^(green|yellow|red)$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Risk"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "pattern": "^(all|pending_confirmation|with_barrier|missed)$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CareTeamPatientRow"
                  },
                  "title": "Response List Patients Api V1 Care Team Patients Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Patients",
        "tags": [
          "equipo asistencial"
        ]
      }
    },
...
    "/api/v1/milestones/{milestone_id}/reschedule": {
      "post": {
        "operationId": "reschedule_milestone_api_v1_milestones__milestone_id__reschedule_post",
        "parameters": [
          {
            "in": "path",
            "name": "milestone_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Milestone Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RescheduleMilestoneRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MilestoneOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Reschedule Milestone",
        "tags": [
          "equipo asistencial"
        ]
      }
    },
...
    "/api/v1/patients/{patient_id}/milestones": {
      "post": {
        "operationId": "create_milestone_api_v1_patients__patient_id__milestones_post",
        "parameters": [
          {
            "in": "path",
            "name": "patient_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Patient Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMilestoneRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MilestoneOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Milestone",
        "tags": [
          "equipo asistencial"
        ]
      }
    },
...
  }
}